ml_service/requirements.txt
# Core ML Libraries
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2

# Multimodal Embeddings
transformers==4.35.0
clip-by-openai==1.0.1
Pillow==10.1.0
sentence-transformers==2.2.2

# Collaborative Filtering
implicit==0.7.2
lightfm==1.17

# Experiment Tracking
mlflow==2.9.0
tensorboard==2.15.1

# API and Serving
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data Processing
pyarrow==14.0.1
datasets==2.15.0
tqdm==4.66.1

# Vector Search (Optional for scaling)
faiss-cpu==1.7.4
annoy==1.17.3

# Evaluation
scikit-surprise==1.1.3
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0

# Docker and Deployment
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
loguru==0.7.2
joblib==1.3.2
requests==2.31.0
